Add MultiheadAttention test, fix dimension calculation
Also added some language to the comments about the sequence lengths.
