+++ src/mlpack/tests/ann/layer_test.cpp
@@ -42,3 +42,4 @@
 #include "layer/softmin.cpp"
 #include "layer/ftswish.cpp"
 #include "layer/layer_norm.cpp"
+#include "layer/multihead_attention.cpp"
